refactor, remove num_neuron constant
